+++ sdk/search/azure-search-documents/src/main/java/com/azure/search/documents/util/SearchPagedFlux.java
@@ -3,6 +3,11 @@
 
 package com.azure.search.documents.util;
 
+import java.util.List;
+import java.util.Map;
+import java.util.function.Function;
+import java.util.function.Supplier;
+
 import com.azure.core.http.rest.PagedFluxBase;
 import com.azure.core.util.paging.ContinuablePagedFlux;
 import com.azure.search.documents.implementation.models.SearchFirstPageResponseWrapper;
@@ -10,19 +15,10 @@ import com.azure.search.documents.implementation.models.SearchRequest;
 import com.azure.search.documents.implementation.util.SemanticSearchResultsAccessHelper;
 import com.azure.search.documents.models.DebugInfo;
 import com.azure.search.documents.models.FacetResult;
-import com.azure.search.documents.models.QueryAnswerResult;
 import com.azure.search.documents.models.SearchResult;
-import com.azure.search.documents.models.SemanticErrorReason;
-import com.azure.search.documents.models.SemanticQueryRewritesResultType;
 import com.azure.search.documents.models.SemanticSearchResults;
-import com.azure.search.documents.models.SemanticSearchResultsType;
 
 import reactor.core.publisher.Mono;
-
-import java.util.List;
-import java.util.Map;
-import java.util.function.Function;
-import java.util.function.Supplier;
 
 /**
  * Implementation of {@link ContinuablePagedFlux} where the continuation token type is {@link SearchRequest}, the
@@ -132,61 +128,4 @@ public final class SearchPagedFlux extends PagedFluxBase<SearchResult, SearchPag
             return Mono.just(metaData.getFirstPageResponse().getDebugInfo());
         });
     }
-
-    /**
-     * The query answers that can be used to further explore your search results.
-     *
-     * @return The query answers that can be used to further explore your search results.
-     */
-    public Mono<List<QueryAnswerResult>> getQueryAnswers() {
-        return metadataSupplier.get().flatMap(metaData -> {
-            if (metaData.getFirstPageResponse().getQueryAnswers() == null) {
-                return Mono.empty();
-            }
-            return Mono.just(metaData.getFirstPageResponse().getQueryAnswers());
-        });
-    }
-
-    /**
-     * The semantic error reason if the semantic search request failed.
-     *
-     * @return The semantic error reason if the semantic search request failed, otherwise an empty Mono.
-     */
-    public Mono<SemanticErrorReason> getSemanticErrorReason() {
-        return metadataSupplier.get().flatMap(metaData -> {
-            if (metaData.getFirstPageResponse().getSemanticErrorReason() == null) {
-                return Mono.empty();
-            }
-            return Mono.just(metaData.getFirstPageResponse().getSemanticErrorReason());
-        });
-    }
-
-    /**
-     * The semantic search results type if semantic search was requested.
-     *
-     * @return The semantic search results type if semantic search was requested, otherwise an empty Mono.
-     */
-    public Mono<SemanticSearchResultsType> getSemanticSearchResultsType() {
-        return metadataSupplier.get().flatMap(metaData -> {
-            if (metaData.getFirstPageResponse().getSemanticSearchResultsType() == null) {
-                return Mono.empty();
-            }
-            return Mono.just(metaData.getFirstPageResponse().getSemanticSearchResultsType());
-        });
-    }
-
-    /**
-     * The semantic query rewrites result type if semantic search was requested.
-     *
-     * @return The semantic query rewrites result type if semantic search was requested, otherwise an empty Mono.
-     */
-    public Mono<SemanticQueryRewritesResultType> getSemanticQueryRewritesResultType() {
-        return metadataSupplier.get().flatMap(metaData -> {
-            if (metaData.getFirstPageResponse().getSemanticQueryRewritesResultType() == null) {
-                return Mono.empty();
-            }
-            return Mono.just(metaData.getFirstPageResponse().getSemanticQueryRewritesResultType());
-        });
-    }
-
 }